redux updates
