+++ src_ts/redux/actions/interventions.ts
@@ -1,21 +1,1 @@
-import {AnyObject} from '../../types/globals';
-import {_sendRequest} from '../../components/pages/interventions/intervention-tab-pages/utils/request-helper';
-// eslint-disable-next-line max-len
-import {Intervention} from '../../components/pages/interventions/intervention-tab-pages/common/models/intervention.types';
-import {etoolsEndpoints} from '../../endpoints/endpoints-list';
-import {getEndpoint} from '../../endpoints/endpoints';
 
-export const setIntervention = (intervention: AnyObject) => {
-  return {
-    type: 'UPDATE_CURRENT_INTERVENTION',
-    current: intervention
-  };
-};
-
-export const getIntervention = (interventionId: string) => (dispatch: any) => {
-  return _sendRequest({
-    endpoint: getEndpoint(etoolsEndpoints.intervention, {interventionId: interventionId})
-  }).then((intervention: Intervention) => {
-    dispatch(setIntervention(intervention));
-  });
-};
